requirements.txt
attrs==25.3.0
Automat==25.4.16
CacheControl==0.14.3
certifi==2025.4.26
charset-normalizer==3.4.2
ConfigArgParse==1.7
constantly==23.10.4
docutils==0.21.2
filelock==3.18.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
lunr==0.7.0.post1
msgpack==1.1.0
mypy==1.16.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pydoctor==24.11.2
pyserial==3.5
pytest==8.3.5
requests==2.32.3
setuptools==80.8.0
Twisted==24.11.0
types-pyserial==3.5.0.20250326
typing_extensions==4.13.2
urllib3==2.4.0
zope.interface==7.2
